// packages/discord/src/operations/createMessage.ts
export const CreateMessageInput = /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
  channel_id: Schema.String.pipe(T.PathParam()),
  content: Schema.optional(Schema.NullOr(Schema.String)),
  embeds: Schema.optional(
    Schema.NullOr(
      Schema.Array(
        Schema.Struct({
          type: Schema.optional(Schema.NullOr(Schema.String)),
          url: Schema.optional(Schema.NullOr(Schema.String)),
          title: Schema.optional(Schema.NullOr(Schema.String)),
          color: Schema.optional(Schema.NullOr(Schema.Number)),
          timestamp: Schema.optional(Schema.NullOr(Schema.String)),
          description: Schema.optional(Schema.NullOr(Schema.String)),
          author: Schema.optional(Schema.Unknown),
          image: Schema.optional(Schema.Unknown),
          thumbnail: Schema.optional(Schema.Unknown),
          footer: Schema.optional(Schema.Unknown),
          fields: Schema.optional(
            Schema.NullOr(
              Schema.Array(
                Schema.Struct({
                  name: Schema.String,
                  value: Schema.String,
                  inline: Schema.optional(Schema.NullOr(Schema.Boolean)),
                }),
              ),
            ),
          ),
          provider: Schema.optional(Schema.Unknown),
          video: Schema.optional(Schema.Unknown),
        }),
      ),
    ),
  ),
  allowed_mentions: Schema.optional(Schema.Unknown),
  sticker_ids: Schema.optional(Schema.NullOr(Schema.Array(Schema.String))),
  components: Schema.optional(Schema.NullOr(Schema.Array(Schema.Unknown))),
  flags: Schema.optional(Schema.NullOr(Schema.Number)),
  attachments: Schema.optional(
    Schema.NullOr(
      Schema.Array(
        Schema.Struct({
          id: Schema.String,
          filename: Schema.optional(Schema.NullOr(Schema.String)),
          description: Schema.optional(Schema.NullOr(Schema.String)),
          duration_secs: Schema.optional(Schema.NullOr(Schema.Number)),
          waveform: Schema.optional(Schema.NullOr(Schema.String)),
          title: Schema.optional(Schema.NullOr(Schema.String)),
          is_remix: Schema.optional(Schema.NullOr(Schema.Boolean)),
        }),
      ),
    ),
  ),
  poll: Schema.optional(Schema.Unknown),
  shared_client_theme: Schema.optional(Schema.Unknown),
  message_reference: Schema.optional(Schema.Unknown),
  nonce: Schema.optional(Schema.Unknown),
  enforce_nonce: Schema.optional(Schema.NullOr(Schema.Boolean)),
  tts: Schema.optional(Schema.NullOr(Schema.Boolean)),
}).pipe(T.Http({ method: "POST", path: "/channels/{channel_id}/messages" }));
export type CreateMessageInput = typeof CreateMessageInput.Type;
// ...
export type CreateMessageOutput = typeof CreateMessageOutput.Type;

// The operation
export const createMessage = /*@__PURE__*/ /*#__PURE__*/ API.make(() => ({
  inputSchema: CreateMessageInput,
  outputSchema: CreateMessageOutput,
  errors: [BadRequest, Forbidden, NotFound] as const,
}));