// packages/discord/src/operations/updateMessage.ts
export const UpdateMessageInput = /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
  channel_id: Schema.String.pipe(T.PathParam()),
  message_id: Schema.String.pipe(T.PathParam()),
  content: Schema.optional(Schema.NullOr(Schema.String)),
  embeds: Schema.optional(
    Schema.NullOr(
      Schema.Array(
        Schema.Struct({
          type: Schema.optional(Schema.NullOr(Schema.String)),
          url: Schema.optional(Schema.NullOr(Schema.String)),
          title: Schema.optional(Schema.NullOr(Schema.String)),
          color: Schema.optional(Schema.NullOr(Schema.Number)),
          timestamp: Schema.optional(Schema.NullOr(Schema.String)),
          description: Schema.optional(Schema.NullOr(Schema.String)),
          author: Schema.optional(Schema.Unknown),
          image: Schema.optional(Schema.Unknown),
          thumbnail: Schema.optional(Schema.Unknown),
          footer: Schema.optional(Schema.Unknown),
          fields: Schema.optional(
            Schema.NullOr(
              Schema.Array(
                Schema.Struct({
                  name: Schema.String,
                  value: Schema.String,
                  inline: Schema.optional(Schema.NullOr(Schema.Boolean)),
                }),
              ),
            ),
          ),
          provider: Schema.optional(Schema.Unknown),
          video: Schema.optional(Schema.Unknown),
        }),
      ),
    ),
  ),
  flags: Schema.optional(Schema.NullOr(Schema.Number)),
  allowed_mentions: Schema.optional(Schema.Unknown),
  sticker_ids: Schema.optional(Schema.NullOr(Schema.Array(Schema.String))),
  components: Schema.optional(Schema.NullOr(Schema.Array(Schema.Unknown))),
  attachments: Schema.optional(
    Schema.NullOr(
      Schema.Array(
        Schema.Struct({
          id: Schema.String,
          filename: Schema.optional(Schema.NullOr(Schema.String)),
          description: Schema.optional(Schema.NullOr(Schema.String)),
          duration_secs: Schema.optional(Schema.NullOr(Schema.Number)),
          waveform: Schema.optional(Schema.NullOr(Schema.String)),
          title: Schema.optional(Schema.NullOr(Schema.String)),
          is_remix: Schema.optional(Schema.NullOr(Schema.Boolean)),
        }),
      ),
    ),
  ),
}).pipe(
  T.Http({
    method: "PATCH",
    path: "/channels/{channel_id}/messages/{message_id}",
  }),
);
export type UpdateMessageInput = typeof UpdateMessageInput.Type;
// ...
export type UpdateMessageOutput = typeof UpdateMessageOutput.Type;

// The operation
export const updateMessage = /*@__PURE__*/ /*#__PURE__*/ API.make(() => ({
  inputSchema: UpdateMessageInput,
  outputSchema: UpdateMessageOutput,
  errors: [BadRequest, Forbidden, NotFound] as const,
}));